# apps/backend/openapi.yaml
openapi: 3.1.0
info:
  title: PowerChain Renewable Miner API
  version: 1.3.0
  description: Multi-tenant control plane for renewable Proof-of-Energy nodes and MINER rewards.
servers:
  - url: http://localhost:3100
paths:
...
  /api/v1/devices/register:
    post:
      summary: Enroll a renewable node
      security: [{deviceApiKey: []}]
      responses: {"201": {description: Device enrolled}}
  /api/v1/proofs:
    get:
      summary: List proofs in an authenticated client scope
      security: [{bearerAuth: []}]
      responses: {"200": {description: Proofs}}
    post:
      summary: Submit signed Proof of Energy for evidence review
      responses:
        "200": {description: Idempotent replay of an already stored proof}
        "202": {description: Proof received and pending attestation}
        "409": {description: Sequence, digest continuity, or source identity conflict}
        "425": {description: Reward/epoch/verification policy is not ready}
        "422": {description: Proof violates physical or timestamp limits}
...
  /api/v1/agents/{agentId}/compute/api-keys:
    post:
      summary: Generate a one-time Agent Compute API key
      security: [{bearerAuth: []}]
      responses: {"201": {description: API key generated}}
  /api/v1/agents/{agentId}/compute/topups:
    post:
      summary: Create a wallet funding intent for compute credit
      security: [{bearerAuth: []}]
      responses: {"201": {description: Top-up intent created}}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    deviceApiKey:
      type: http
      scheme: bearer